+++ proknow/Patients/Entities/ImageSets.py
@@ -1,4 +1,6 @@
 import os
+import time
+import datetime
 
 from .EntityItem import EntityItem
 from ...Exceptions import InvalidPathError
@@ -67,6 +69,53 @@ class ImageSetItem(EntityItem):
             image_path = os.path.join(main_directory, modality + "." + image["uid"])
             self._requestor.stream('/workspaces/' + self._workspace_id + '/imagesets/' + self._id + '/images/' + image["id"] + '/dicom', image_path)
         return main_directory
+
+    def get_image_data(self, index):
+        """Gets the image data for the image at the given index.
+
+        Parameters:
+            index (int): The index of the image for which to get the data.
+
+        Returns:
+            bytes: The image data.
+
+        Raises:
+            AssertionError: If the input parameters are invalid.
+            :class:`proknow.Exceptions.HttpError`: If the HTTP request generated an error.
+
+        Example:
+            This example shows how to get the image data for each image in an image set::
+
+                from proknow import ProKnow
+
+                pk = ProKnow('https://example.proknow.com', credentials_file="./credentials.json")
+                patients = pk.patients.lookup("Clinical", ["HNC-0522c0009"])
+                patient = patients[0].get()
+                entities = patient.find_entities(type="image_set")
+                image_set = entities[0].get()
+                slice_count = len(image_set.data["data"]["dicom"])
+                slice_data = [image_set.get_image_data(i) for i in range(slice_count)]
+        """
+        assert isinstance(index, int), "`index` is required as an integer."
+        headers = {
+            "Accept-Version": "1",
+            "Authorization": 'Bearer ' + self._data["data"]["dicom_token"]
+        }
+        start = datetime.datetime.now()
+        DELAY = 0.2
+        while (datetime.datetime.now() - start).total_seconds() < self._proknow.ENTITY_WAIT_TIMEOUT:
+            _, imageset = self._rtv.post('/imageset', json={"data": self._data["data"]["dicom"]}, headers=headers)
+            if imageset["status"] == "completed":
+                break
+            else: # pragma: no cover (unreliable)
+                time.sleep(DELAY)
+        else: # pragma: no cover (unlikely)
+            pass
+        image = imageset["data"]["images"][index]
+        pid = imageset["data"]["processed_id"]
+        iid = image["processed_id"]
+        _, content = self._rtv.get_binary('/imageset/' + pid + '/image/' + iid, headers=headers)
+        return content
 
     def refresh(self):
         """Refreshes the image set entity.